+++ libs/base_core/src/sys/mem/size/expr/mod.rs
@@ -1,4 +1,4 @@
-// devela::sys::mem::size::expr
+// devela_base_core::sys::mem::size::expr
 //
 // Original source code by Joshua Nelson, licensed as BSD-3,
 // https://crates.io/crates/size-of-trait/1.1.3
@@ -16,7 +16,7 @@ pub const fn __size_of_expr<T>(_zero_len_fn_ptr_array: [impl FnOnce() -> [T; 0];
 ///
 /// # Example
 /// ```
-/// # use devela::size_of_expr;
+/// # use devela_base_core::size_of_expr;
 /// async fn f() {
 ///     let x = 1;
 ///     core::future::ready(()).await;
@@ -59,7 +59,7 @@ pub use size_of_expr;
 /* tests */
 
 // has to be a separate file because of experimental syntax
-#[cfg(all(test, nightly_coro, feature = "alloc"))]
+#[cfg(all(test, nightly_coro))]
 mod test_coro;
 
 #[cfg(test)]